예약기능 구현 시작

+++ src/main/java/kr/or/iei/tour/model/vo/TourVO.java
@@ -26,5 +26,5 @@ public class TourVO {
 	private int itemPrice;
 	private String tourTimes;
 	private int readCount;
-	private float score;
+	private double score;
 }
